requirements.txt

fastapi
uvicorn
jinja2
python-multipart
python-dotenv
requests
azure.cognitiveservices.speech
azure-storage-blob
openai